label_map 取得部を関数化
